add carts

+++ lib/models/login_model.dart
@@ -20,7 +20,7 @@ class UserData
   String phone;
   String image;
   int points;
-  int credit;
+  double credit;
   String token;
 
   UserData.fromJson(Map<String, dynamic> json)
